增加htmlpaste和包大小分析

+++ src/src/editor/plugins/MarkDownShortcuts/onSpace.js
@@ -15,18 +15,19 @@ export default function onSpace(event, editor, next) {
 	if (type) {
 		//如果类型匹配为heading则址
 		console.log(chars, type);
-
 		if (type.match(/heading/) || type.match('paragraph')) {
-			// 防止空格
-			event.preventDefault();
-		}
-		if (startBlock.type.match(/heading/) && !type.match(/heading/)) {
-			return next();
+			if (startBlock.type !== 'list-item') {
+				// 命中如果不是列表则防止空格
+				event.preventDefault();
+			}
+			if (startBlock.type.match(/heading/) && !type.match(/heading/)) {
+				return next();
+			}
 		}
 		if (type === 'list-item') {
 			return editor.insertListItem(); //不传入参数，则由正则判断，传入字符串，表示类型	 'ordered' ｜ 'undo'｜'finished'｜ 'bulleted'
 		}
-		return editor.moveFocusToStartOfNode(startBlock).delete().setBlocks(type);
+		if (startBlock.type !== 'list-item') return editor.moveFocusToStartOfNode(startBlock).delete().setBlocks(type);
 	}
 
 	editor.withoutNormalizing((editor) => {
